Add elixir_make workaround
If we don't set ELIXIR_MAKE_CACHE_DIR, the elixir_make package attempts
to write to a cache directory under $HOME, which fails.

Get many of the elixir_make items working by adding this and adding
gnumake as a build input.

For some elixir_make packages, they'll still need additional
depenencies, which can be added on a package name basis.

Note: It feels like this and rustlerPrecompiled, which both examine
dependencies to decide on workarounds could potentially be done purely
in nix instead of examining the dependencies in the deps_nix code. Doing
that would also make it more clear to others how to add local
workarounds for all crates that depend on a particular crate. I've kept
this in deps_nix's elixir code for now for consistency with
rustlerPrecompiled.